requirements.txt
torch --index-url https://download.pytorch.org/whl/cu118
torchtext
accelerate
tqdm
pandas
gensim
torchvision
torchinfo
dash
dash_canvas
opencv-python
numpy==1.23.1
matplotlib
